Update `Too many errors for endpoint` message from Error to Warn (#42097)
### What does this PR do?

Resolves #39921.

Updates the error message:

```
Too many errors for endpoint '*': retrying later
```

### Motivation

This message is just emitted when the forwarder temporarily suppresses sending transactions to an endpoint that has recently failed, in order to avoid flooding it whilst in an error state. It is not an error in itself.

### Describe how you validated your changes

Tested manually by running against a dummy endpoint. Killed the endpoint and ensured the messages were now a warning.

### Additional Notes