Define concrete Template in ir program
This defines the structure of a template as relevant to an ir.Program

The individual segment kinds (StringSegment and JSONSegment) hold not
just the definitions of user defined data to evaluate but expression
kind and index to tie this template segment into the program expressions.

The newly defined RootExpressionKindTemplateSegment is not particularly
useful as part of this set of commits but will be relevant when implementing
future operations.