Revert "Revert "[ACIX-835] Update testify to v1.11.1" (#42017)" (#42020)
This reverts commit aa9e94fba263401c21434516418a98da1bb98bcf.

### What does this PR do?

Re-apply #41981 and fix the tests that it broke.

### Motivation

Same as #41981: speed up tests execution thanks to stretchr/testify#1427 that makes `require.EventuallyWithT` test the condition immediately instead of waiting for an initial delay before making the very first try.

### Describe how you validated your changes

Launched a [full pipeline](https://gitlab.ddbuild.io/DataDog/datadog-agent/-/pipelines/79481153) to be sure to run all the tests and not a subset of them: https://github.com/DataDog/datadog-agent/pull/42020#issuecomment-3410510955

### Additional Notes

The main feature we were looking for by upgrading `testify` to `v1.11.1` was the fact that that `require.EventuallyWithT` is now evaluating the condition the first time immediately instead of waiting for a `tick` delay before making the first try.
This unexpectedly broke the tests that are touched in this PR.
Fixing tests by adding a `time.Sleep(…)` is most probably not the smartest way of fixing them. But it restores the execution time flow of the previous versions of `testify` that happened to make the tests pass.

Co-authored-by: guyarb <guy20495@gmail.com>