Replace sender manager with noop in diagnose commands
Checks running from flare and diagnose commands are not expected to send
metrics to the backend and does not need to have access to the metrics
pipeline, to which the sender manager is the entrypoint.

Replace diagnoseSenderManager with noop sender manager, that will
return an error if a check attempts to send a metric.