dyninst/symdb: don't yield duplicate packages
Have the packages iterator keep track of which packages it already
yielded and ignore compile units belonging to already-yielded packages.
Go seems to produce multiple compile units for packages that include
assembly source files -- one unit for the .go files, and a unit for each
.s file. Before this patch, the iterator was returning such duplicates,
and we were uploading and ingesting them. This was mostly fine (albeit
confusing) as far as the backend was concerned, but now I'm changing the
backend to overwrite the symbols from a package that already exists,
rather than storing them side by side.

This patch makes it so that we only yield the symbols from the first
compile unit for each package name, which empirically corresponds to the
.go files.