dyninst/symdb: tag package symbols with agent version
Tag each package scope that we upload with the version of the agent that
produced the symbol information. I'm using a new key in
language-specifics for package scopes.
This information seems generally useful for debugging, but also I
intend to use it on the backend to decide if new symbols being uploaded
should overwrite existing ones or, if they're uploaded by the same agent
version as before, should be ignore.

The agent version is technically available to the backend even without
the agent explicitly including this info in the upload -- we have the
agent version as a tag on the log message carrying the symbols
attachment, so the backend could read it from there and add it to the
symbols, or persist it somewhere else. Still, including it explicitly in
the upload seems cleaner.