[SQG] Measure docker image size in container_build jobs (#41812)
### What does this PR do?
This is **_PART II_**, the implementation is not complete yet and not used productively. However to not spawn another PR containing dozens of changes it was decided to bring this work to main step by step.

* Add `!reference [.static_quality_gate_docker]` to the end of the script block of `.docker_build_job_definition` to analyze produced image. Unfortunately, due to several limitations it's not an in place measurement as we use crane that has to pull from remote registry each time.
* Create visual reports in the pipeline logs and save a full yaml along with the full file inventory to Gitlab artifacts (and S3 bucket).
* It is also possible to run analysis locally. Example command, it is important to set the correct gate name that exists in test/static/static_quality_gates.yml:

```bash
dda inv quality-gates.measure-image-local \
    --image-ref "registry.ddbuild.io/ci/datadog-agent/dogstatsd:v74596342-e431809e-amd64" \
    --gate-name "static_quality_gate_docker_cluster_agent_amd64" \
    --build-job-name "local" \
    --output-path "report.yml" \
    --debug

```
### Motivation
We want to provide as much information to the developers as possible and also ensure good performance. It is painful right now and mostly reactive work as developers start to respect Static Quality Gates only after they become red in main.