process collector: Use check client for discovery
The process collector currently uses the normal system probe client when
talking to the discovery module but that client's timeout (5 seconds) is
too short for the discovery services endpoint, leading to spurious
timeouts.  Also, the process collector duplicates the "emit warnings
during system-probe startup" logic in the system probe check client.

Use the system probe check client from the process collector to fix
this.  Since the services endpoint needs a request body, also add a
helper function in the check client to send requests with bodies.