Maxime/cleanup/migrate helper func out of setup (#41891)
### What does this PR do?

This is a cleanup PR which includes no change in behavior. The goal is to clean `pkg/config/setup` from the myriad of helpers that are not related to configuration setup. Some were migrated to `pkg/config/utils`. Others were migrated within the code they relate to as they're only used once and extremely tied to their product behavior.

**This PR is better reviewed commit by commit**:

This includes:
- Migrate `GetRemoteConfigurationAllowedIntegrations` to `comp/core/autodiscovery/providers`
- Move `IsAgentTelemetryEnabled` from config setup to utils
- Extract `GetObsPipelineURL` logic to `comp/forwarder/defaultforwarder` (the only place where it's used)
- Remove `GetValidHostAliases` from config setup
- Make `ResolveSecrets` and `EnvVarAreSetAndNotEqual` method private
- Move GetBindHost from `setup` to `utils`
- Move IsCloudProviderEnabled from `setup` to `utils`

### Motivation

We want the `pkg/config/setup` package to only contain initialization logic to reduce package dependency and simplify the config initialization.

### Describe how you validated your changes

Running the CI is enough.