Fix Windows Fleet Setup Script (#42085)
### What does this PR do?
Fixes a bug in the windows installer to not try and install the non exsistient datadog-inject package and fail install when IIS options are set. 

### Motivation
https://datadoghq.atlassian.net/browse/WINA-1930
https://datadoghq.atlassian.net/browse/WINA-1356

### Describe how you validated your changes

Adds e2e test that uses install script to install agent with APM IIS SSI options. 

### Additional Notes
When host option added for IIS we should better split this logic to install correct package.