Build popt with bazel (#42139)
### What does this PR do?

Build popt with bazel from omnibus

### Motivation

Migrating our native dependencies to bazel

### Describe how you validated your changes

Local build (and CI when AWS recovers)

### Additional Notes

This local_define: `'POPT_SYSCONFDIR=\\"/etc/\\"',` is open to debate. 
In theory it is derived from the install location, but needs to be known at build time, which we can't do when building with bazel.
The value is used to load potential `popt` configuration files.
The current location is `/opt/datadog-agent/etc/` which doesn't exist, so any configuration loading will fail, while it could succeed with the new value, but I'm not sure if we want it to potentially succeed.

We could also ensure it never succeed by passing an ostensibly invalid path.