[clusteragent/admission] Decouple default label selectors from APM opts (#41977)
### What does this PR do?

This PR moves APM-specific config options out of the default label selectors used in the DCA admission controller.

The default label selectors are used by 4 webhooks: APM, config, tags-from-labels, and kubernetes-admission-events. In the past, the config and tags-from-labels webhooks needed to run when APM was enabled. They're now decoupled, so this is no longer needed. The kubernetes-admission-events webhook was tied to APM by accident because it has nothing to do with APM.

This PR moves the current default label selectors to the APM webhook without changes and modifies the default ones used by the other 3 webhooks to avoid depending on APM config options.

Not sure if kubernetes-admission-events should check `mutate_unlabelled` since it validates rather than mutates. I'll keep the existing behavior for now.

The APM label selector logic is unchanged. There may be other optimization opportunities, but I'm leaving those for injection-platform.

### Describe how you validated your changes

CI and manual tests to verify that basic injection scenarios work as expected.


Co-authored-by: david.ortiz <david.ortiz@datadoghq.com>