Notify tracer metadata memfd creation via eBPF (#41742)
### What does this PR do?

There is code in the agent to get tags from the tracer metadata of
processes and attach them to network connections by using the tagger
component.  However, in the case of short-lived processes, or during
process startup, network data may be available before the process'
metadata is available in the agent. To significantly reduce the race
window, add an eBPF event to monitor the sealing of the memfd used by
the tracers, and read the memfd when that event is received.

We hook memfd_create(2) to check that the name of the memfd starts with
the prefix we want.  We store the fd and the 8-byte suffix of the memfd
name in a map.  In memfd_fcntl(), we check if the memfd we want is being
sealed (indicating that the tracer has finished writing it), and report
that to user space.

While this should solve many cases, note that there are some cases
that we don't handle in this PR:
- If the process exits even before we are able to read the memfd from userspace. This will likely require changes to the tracers, to use eRPC or something similar, so will be considered separately in the future.
- Existing processes. These are already handled by the code in the agent which scans all existing processes at startup and puts the tags in the tagger. In the future we can consider adding something similar to the /proc-based event generation which the process event code has.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-197

### Describe how you validated your changes

Tests added.

### More information

[Specification for tracer metadata](https://docs.google.com/document/d/1kcW6BLdYxXeTSUz31cBqoqfW1Jjs0IDljfKeUfIRQp4/edit?tab=t.0#heading=h.8d3o7vtyu1y1)
[PR in core agent to use tagger in connections check](https://github.com/DataDog/datadog-agent/pull/41347)
[PR in core agent to add tracer metadata tags to tagger](https://github.com/DataDog/datadog-agent/pull/41306)




Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>