Move logger formats to pkg/util/log  (#42177)
### What does this PR do?
Unexport `LoggerFromWriterWithMinLevelAndFormat`, and export functions for each format instead. 

### Motivation
The current log format string is very `seelog` specific, so when we change logger we'll have to change the format string (not the log output though), which would result in many conflicts if the PR takes time.

### Describe how you validated your changes
CI.
This only impacts tests.

### Additional Notes
For context: opened https://github.com/DataDog/datadog-agent/pull/39139 a few months ago and now it has conflicts everywhere.

Co-authored-by: pierre.gimalac <pierre.gimalac@datadoghq.com>