[INCIDENT-44715] Checkout code from correct base branch in APM benchmark job (#42267)
### What does this PR do?

Dynamically determine what the correct base branch is using the
`release.json` file, and use that in the `benchmark` CI job instead of
relying on a hardcoded `main`.

### Motivation

incident-44715
The APM benchmark job always checks out `main`, which is not always the
correct base branch - for example when running in a PR on a release
branch.

This can cause issues, for example in the current incident where the
`dda` version required on `main` exceeds the version installed for
release branch CI pipelines.

### Describe how you validated your changes

Running CI and checking logs.

### Additional Notes