dyninst: refactor process monitoring abstractions (#42302)
### What does this PR do?

This is a completely non-functional change. It adds a layer of abstraction to the `module` package for getting notified of changes to processes and their configs: `ProcessSubscriber`.

The plumbing between process monitoring (procmon) and remote configuration scraping (rcscrape) used to be handled explicitly in the module package, which was in the way of my introducing a fresh implementation of this sort of logic.

* Extract shared process types into `pkg/dyninst/process` package. The use of types in `rcscrape` and `procmon` was chaotic.
  - Info: process instrumentation metadata
  - ID: process identifier
  - Config: instrumentation configuration including probes and runtime
  - ProcessesUpdate: aggregated removals and configuration updates

* Create `pkg/dyninst/procscrape` package with `Subscriber` type that bridges procmon and rcscrape, coalescing process lifecycle events and remote-config updates into a unified stream.

* Move process.go → process/info.go and add ContainerInfo, GitInfo types to the shared package

* Update all consumers (module, actuator, rcscrape, procmon, uprobe) to use the new shared types from pkg/dyninst/process

### Motivation

This is handy cleanup to make room for a fresh implementation of process subscribing
via the implementation of the [implicit enablement rfc](https://docs.google.com/document/d/1Krv5mdRiPxY0b-HsJ0vGn8GnUcll0RM4__2-SIdLcEw/edit?tab=t.0).

### Describe how you validated your changes

No functional changes. All the tests pass.

### Additional Notes

Part of  [DEBUG-4589](https://datadoghq.atlassian.net/browse/DEBUG-4589).

[DEBUG-4589]: https://datadoghq.atlassian.net/browse/DEBUG-4589?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>