Consolidate part of the domains configuration for the forwarder
Move bits of logic split between the component (forwarder.go) and the
package (defaultforwarder.go) into the resolver package, so it can
provide a more unified front.

Now defaultforwarder-the-library always gets the resolvers ready to
use, and it is responsibility of the callers to provide
one. Everything was doing that already, except the
defaultforwarder-the-component.