APM: no panic on forwarding transport errors (#42339)
### What does this PR do?
The http roundtrip transport can return a nil response for certain kinds of errors. When this happens our attempt to close the response body results in a nil pointer exception. We don't want to crash.

### Motivation
panicking is no good for software (and people too probably)

### Describe how you validated your changes
New unit tests (Written with AI assistance but reviewed by me (a human))

### Additional Notes


Co-authored-by: andrew.glaude <andrew.glaude@datadoghq.com>