[AGENTRUN-516] Check dead code elimination enabled in agent binaries in deploy pipelines (#38366)
<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?
Check whether Go binaries have dead code elimination enabled when building in deploy pipelines.
If the optimization is disabled, the check just logs it (with info on why it's disabled), but doesn't fail the build.

### Motivation
Provides a way to know whether the optimization is enabled in the various binaries that we build, automatically at regular interval (nightlies) and manually if anyone wants to.
This can help debug size increases if you're not familiar with running this locally (or on other platforms).

Only running in deploy pipelines as it increases the duration of the build job by a few minutes.

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes. If you validated changes manually and not through automated tests, add context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask reviewers to do it. Describe how to set up an environment for manual tests in the PR description. Manual validation is expected to happen on every commit before merge.

Any manual validation step should then map to an automated test. Manual validation should not substitute automation, minus exceptions not supported by test tooling yet.
-->
CI.
Lots of manual testing.

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->


Co-authored-by: pierre.gimalac <pierre.gimalac@datadoghq.com>