(fleet) fix windows config experiments
This PR re-introduces windows config experiments, tested against the 3 standard failure modes:
- Timeout
- Crash
- Backend says stop experiment because it thinks the agent is unhealthy

Similarly to the previous implementation the rollback of the experiment assumes no issue with the background process. Reboots, crashes, etc... will prevent the rollback from occuring. Should be fixed in a subsequent PR but requires designing a way to only use the experiment config for one restart.