discovery: Fix and enable node test (#42357)
### What does this PR do?

The test server includes dd-trace in its package.json, but it doesn't
actually initialize the tracer (and besides that it uses an old version
of the tracer without metadata information), so it's not correct to
check for instrumentation = true.

We don't want to make the test service use dd-trace since the purpose of
the test server is really something else (USM TLS test).  Tracer
metadata handling is already tested in other unit tests and e2e tests.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-234


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>