[EBPF-882] telemetry: disable per-cpu perf buffer telemetry by default (#42362)
### What does this PR do?

This PR changes the eBPF telemetry so that it can send data perf buffer data per-CPU or aggregated. This is controlled by a configuration setting, disabled by default.

### Motivation

Splitting metrics per-cpu can generate a lot of metrics, specially in nodes with a high processor count. Reducing this number helps reduce the cost of these metrics, and also solves an issue where in some nodes with hundres of cores the agent_telemetry check was reaching the limit of 2000 metrics, with perfbuffer related metrics alone representing more than 2000 contexts.

### Describe how you validated your changes

Ran locally with both configurations and validated that the telemetry works as expected.

### Additional Notes


Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>