[SYNTH-22692]: Set nextRun when receiving a test config with a newer version (#42463)
### What does this PR do?

If a test is updated, it should be run as soon as possible by the Agent.
Currently, we don't set it and it relies on the scheduling it has in-memory.

### Motivation

### Describe how you validated your changes

Validated it locally with the agent running + unit tests.

### Additional Notes

Co-authored-by: rafael.oliveira <rafael.oliveira@datadoghq.com>
(cherry picked from commit 7e46ca85715f5a5d3ec19fb534f7958a3be6552d)