[CWS] allow enabling sysctl snapshot even if the event is disabled (#42482)
### What does this PR do?

This PR creates a new config option managing the emission of the sysctl ebpf event. Now the main config controls both the ebpf event and the snapshot event, allowing to disable sysctl fully like today. The win is that now you can enable the snapshot event without the ebpf event being supported or enabled (as is the case on cgroup v1 hosts).

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>