(fleet) Integrate IIS instrumentation with the APM instrument commands (#42086)
### What does this PR do?

This PR does two things:

1. Decouple IIS instrumentation from the dotnet library OCI installation. 

IIS instrumentation is only turned on if `DD_APM_INSTRUMENTATION_ENABLED=iis` is set.
One consequence of this is that IIS instrumentation will not be updated when the dotnet library package is remotely updated. Remote library upgrades are not released, so we'll only support remote library upgrades for the host level instrumentation. I have removed two tests that were exercising this path.

2. Integrate IIS instrumentation with the `apm instrument` and `apm uninstrument` commands and update the `iis-instrumentation.bat` script accordingly.


### Motivation

### Describe how you validated your changes

E2E tests

### Additional Notes


Co-authored-by: bertrand.mermet <bertrand.mermet@datadoghq.com>