[ AGNTLOG-310 ] Lock down dynamic symbols exported in the agent binary (#41049)
### What does this PR do?
The goal of this PR is to restrict the exported dynamic symbols exported by the agent to only include those necessary for runtime operations. A go dependency (go-nvml) included in some build versions of the agent has triggered an automatic export of all symbols within the binary, which has in turn caused some symbol conflicts/segfaults that needed significant investigation to identify and fix. While the individual error case has been patched, removing unneeded dynamic symbols offers a much more secure mitigation moving forward.

Note that we *DO* export the nvml symbols as global symbols in this change - this is a tentative choice intended to be reviewed here. All of the existing test setups for the subset of functionality that interacts with nvml report successful runs with symbols exported. 

### Motivation

### Describe how you validated your changes
Ultimately the testing of this change was reliant on the existing suite of automated tests. The binary was hand verified as fully functional for the purposes of the logs agent, but no other system was manually vetted.


### Additional Notes


Co-authored-by: ryan.hall <ryan.hall@datadoghq.com>