trace/api: fix forwardingTransport with multiple targets (#42515)
### What does this PR do?

Before this change we'd close the response body for the request that we RoundTrip which violates the contract of the RoundTripper used by the ReverseProxy. The result was that all the proxied requests looked like they had failed. In practice they hadn't, but it lead to quite a bit of scary logging, and it prevented the client of the proxy from actually getting the response.

Also, limit the rate of error logging when sending to secondary endpoints.

### Motivation

We recently started using multiple endpoints in some clusters and observed a lot of these logs (this code used to never be exercised). Worse yet, this had caused panics (https://github.com/DataDog/datadog-agent/pull/42339). 

### Describe how you validated your changes

Testing.



Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>