discovery: Remove TestServicesPythonFromBashScript (#42520)
### What does this PR do?

This test jumps through some hoops to find the exact PID we expect and
was flaky due to attempting to reading a command line while a process
was being execed.

But this test should not be needed at all since executing a Python
server from a bash script is no different from executing it from
anywhere else.  So just remove it.

### Motivation

https://datadoghq.atlassian.net/browse/DSCVR-210


Co-authored-by: vincent.whitchurch <vincent.whitchurch@datadoghq.com>