dyninst/decode: allow the last data item to fail (#42530)
### What does this PR do?

We see sometimes that we fail to decode data items. We have theories that it relates to running out of buffer space, but we haven't fully tracked down that bug. In the meantime, let's change the behavior to move along rather than fail hard.

### Motivation

We see this in the wild.

### Describe how you validated your changes

testing

Relates to https://datadoghq.atlassian.net/browse/DEBUG-4569

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>