Obfuscate db.statement and db.query.text
These fields _should_ match the contents of the span resource name, but
this semantic recommendation is not guaranteed. For libraries that do
not follow it, this commit checks whether the values match, before
assigning the value of the obfuscated resource name in the OTel db
fields. If they do not match, a separate obfuscation is performed.