dyninst: protect against repeated crashes when placing probes (#42489)
This patch adds some protection against the system-probe crashing while generating and loading a BPF program for Dynamic Instrumentation probes (for example because of OOM) and then restarting and crashing again while placing the same probe. This is done with a simple mechanism: when starting to load a program, a "tombstone" file is written (/opt/datadog-agent/run/debugger_probes_tombstone.json). If loading finishes (successfully or not), we delete the file. If it doesn't finish, then the restarted process assumes that the previous incarnation crashed and waits for a while before starting the dynamic instrumentation module. If we crash again, the next wait is even longer. Waits start at a minute and go up to an hour. If the system-probe starts with a tombstone file and then manages to stay alive for one minute, the tombstone is deleted.

https://datadoghq.atlassian.net/browse/DEBUG-4601

Co-authored-by: andrew.werner <andrew.werner@datadoghq.com>