[ddot] byoc: update dockerfile (#42233)
### What does this PR do?
This PR updated the BYOC dockerfile in the release branch to point to
the correct stable references for the branch.

### Motivation
Currently customers need to jump through additional hoops and may
inadvertently build an unstable image.

### Describe how you validated your changes
Following the steps described
[here](https://docs.datadoghq.com/opentelemetry/setup/ddot_collector/custom_components/),
but using the correct docker and manifest file references (documentation
to be updated accordingly). Just pull the dockerfile changed here, and
the manifest file in this branch and run:

```
docker build . -t agent-ddot --no-cache
```

The build will succeed and be built on top of the correct image.

### Additional Notes