Add a dependency between fakeintake and VM to avoid rogue agent (#42104)
### What does this PR do?

Add a dependency between the fakeintake and the VM/Cluster where the agent is running.
That dependency is not strictly needed but will allow us to make sure that the Agent is destroyed before we destroy the fakeintake. Without that we can have some scenarios where the agent is still alive, the fakeintake destroyed. And the agent can start polluting other tests that reuse the IP for their fakeintake.

Some details about investigation: https://datadoghq.atlassian.net/wiki/x/HgOzTgE

That's a first step, as a second step I will add the same dependency in EKS or ECS provisioners, but it first need a PR On test-infra side. So starting with that one

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>