fix(otel-agent): Obfuscate db.statement and db.query.text (#42539)
### What does this PR do?
These fields _should_ match the contents of the span resource name, but this semantic recommendation is not guaranteed. 

For libraries that do not follow it, this commit checks whether the values match, before assigning the value of the obfuscated resource name in the OTel db fields. If they do not match, a separate obfuscation is performed.

### Motivation

A recent investigation on a customer escalation showed that `spanner` library does not follow this convention, which results on the trace-agent overriding the contents of `db.statement`.

### Describe how you validated your changes
Added unit tests

### Additional Notes


Co-authored-by: inigo.lopezdeheredia <inigo.lopezdeheredia@datadoghq.com>