DDOT BYOC Dockerfile: take a single version argument
The Dockerfile.agent-otel is meant to be used by end-clients.
By default, only a single version argument is to be provided, e.g. '7.71.2' or '7.73.0-rc1'

Advanced users might want to use a Docker image different from the 7.x.y-full one,
in which case the AGENT_DOCKER_TAG can be overridden.

Also, use the non-dev docker image by default, for end-users not to have to worry about that