[CWS] implement kprobe based fallback for `sched_process_fork` tracepoint  (#42543)
On kernels < 4.15 you can only attach one eBPF probe per tracepoint. This can create conflicts with other programs, especially regarding `sched_process_fork` which is a tracepoint commonly used.

This PR adds a fallback for this tracepoint using kprobes, and uses this fallback on kernel < 4.15.

We have a 4.14 kernel in the CI.

Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>