Make bazel CI templates less error-prone (#42502)
### What does this PR do?
This refactors bazel CI job templates to use centralized
platform-specific runner definitions (`.bazel:runner:*`) instead of
duplicating configuration across multiple jobs.
    
### Motivation
This makes templates less error-prone because:
- platform configuration (`tags`, `extends`) is now defined once in
`defs.yaml` and changes to runner configuration only need to be made in
one place, which reduces copy-paste errors when creating new bazel jobs,
- same applies to the Windows `ARCH` variable. It still looks redundant
with the image we inherit from the default template, but this will be
addressed separately,
- renaming the Windows `SCRIPT` variable (mandated by the peculiar
Docker-in-Docker setup) to `POWERSHELL_SCRIPT` makes it clear it's not
to be confused with GitLab's `script`.

Co-authored-by: Régis Desgroppes <regis.desgroppes@datadoghq.comx>