Make sure we fail on required coverage failure (#42508)
### What does this PR do?

Following #incident-44717, we realized that some coverage computation could be silently failing on the nightly coverage pipeline, leading to incomplete coverage report and an incomplete index then used to determine which test should be skipped.
That PR introduce a mechanism so that a failure to retrieve the coverage on a required agent binary will not make the job fails. So that we are aware that something is not working as expected and so that the dynamic test index is not updated when that happens.
I also disabled the coverage computation on some tests where it was already not working for reasons. Now that the coverage computation is required for jobs that support it we need to explicitly disable it some job that does not support it.

### Motivation

Make dynamic test index computation reliable

### Describe how you validated your changes

Ran a coverage pipeline with the changes

### Additional Notes


Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>