[DDOT][OTAGENT-559] COAT: Add GW usage metric  (#42560)
### What does this PR do?
This PR adds COAT GW usage metric to DDOT.

It relies on the previous non-COAT metric to determine if a Collector is acting as a GW, processing traffic from multiple hosts in the same collector.

### Motivation
The motivation for this feature is to be able to more closely track GW usage adoption and deployments. 

### Describe how you validated your changes
These changes can be easily validated; COAT metric should be reported to the corresponding org for running collectors.

Also added the relevant unit tests to validate the COAT metric values were expected.

### Additional Notes

The current pattern to do/add COAT metrics is not the cleanest, we probably want to work a little bit on the interfaces and abstractions so that it becomes easier and more ergonomic to add new metrics, and, in particular, to report effectively telemetry as both COAT and non-COAT when required. We can tackle that in a separate effort. 


Co-authored-by: jaime.fullaondo <jaime.fullaondo@datadoghq.com>