system-probe: Enable discovery if SP is enabled
Currently, discovery is enabled automatically if USM is enabled, in
order to get process tags from the tracer metadata.  For the same
reason, we will also need to enable discovery if CNM is enabled.
Besides this, there is a stated goal of enabling discovery by default in
general in the future.  As a step towards this, if any system-probe
module is enabled, enable the discovery module too, unless it is
explicitly disabled.